+++ NEWS.md
@@ -2,7 +2,25 @@
 
 Newest versions at top.
 
-## tracerer 2.2
+## tracerer 2.2.1 (unreleased)
+
+### NEW FEATURES
+
+  * None
+
+### MINOR IMPROVEMENTS
+
+  * `save_beast_estimates` can save in sub-sub-folder
+
+### BUG FIXES
+
+  * None
+
+### DEPRECATED AND DEFUNCT
+
+  * None
+
+## tracerer 2.2 (2021-05-14)
 
 ### NEW FEATURES